version 1.0.1
Typo fix that caused the add_rtl_to_p function not to work correctly. this function
is using DOMDocument Class to change the HTML and add the dir=rtl and the text-align=right.
If DOMDocument is not existing a regex is initiated. However, during tests I left behind call to bad name class, so DOMDocument was never called. Now it is called.




git-svn-id: https://plugins.svn.wordpress.org/rtl-rss-feed-fixer/trunk@1027444 b8457f37-d9ea-0310-8a92-e5e31aec5664